bug(fix): Fix admin request success message and move requestMock file
[Finishes #161291958]
